requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
PyYAML>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
transformers>=4.30.0
timm>=0.9.0
